Provide AST default values and support fractional frame rates in Lottie parser
Add sensible defaults across AST models in format/ to prevent MissingFieldException runtime crashes when decoding sparse or partial Lottie JSON files, and change Animation.frameRate, startFrame, and endFrame to Float to support non-integer framerates (e.g. 23.976, 29.97, 59.94).

- Animation: Provide defaults for all fields and support Float framerate and frame spans.
- SolidColorLayer: Provide defaults for solidColor ("#000000"), solidWidth (0f), and solidHeight (0f).
- Fill & Stroke: Provide default StaticColorProperty(Color.Black.rc) for omitted color properties.
- Path: Provide default StaticBezierProperty(BezierValue()) for omitted shape geometry properties.
- Added comprehensive unit tests in LottieDecoderResilienceTest verifying missing-field resilience and fractional framerate decoding.

+++ remotecompose/lottie/src/main/java/com/google/android/horologist/remotecompose/lottie/LottieAnimation.kt
@@ -121,8 +121,8 @@ internal fun LottieAnimation(
   progress: RemoteFloat? = null,
 ) {
   // Total span of frames across the animation timeline.
-  val totalFrames = (animation.endFrame - animation.startFrame).toFloat()
-  val startFrameRf = animation.startFrame.toFloat().rf
+  val totalFrames = animation.endFrame - animation.startFrame
+  val startFrameRf = animation.startFrame.rf
 
   // Determine current frame: if progress [0.0, 1.0] is provided (e.g. via a named variable
   // or user-driven state), map it directly to frames; otherwise drive it continuously
@@ -131,8 +131,7 @@ internal fun LottieAnimation(
     if (progress != null) {
       startFrameRf + (progress * totalFrames)
     } else {
-      startFrameRf +
-        (floor(RemoteFloat(ANIMATION_TIME) * animation.frameRate.toFloat()) % totalFrames)
+      startFrameRf + (floor(RemoteFloat(ANIMATION_TIME) * animation.frameRate) % totalFrames)
     }
   val animationSettings =
     LottieSettings(
@@ -140,7 +139,7 @@ internal fun LottieAnimation(
       slotMap = slotMap,
       width = animation.width.toFloat(),
       height = animation.height.toFloat(),
-      endFrame = animation.endFrame.toFloat(),
+      endFrame = animation.endFrame,
     )
 
   CompositionLocalProvider(LocalAnimationSettings provides animationSettings) {
